move dataclasses to typing.py
